initial commit for the add-on project
The goal of this repository is to play with the go plugin package,
and see how we could load dynamically some function to execute.
That could be useful for a distributed process engine, or
distributed workflow engine.

This is a work in progress, currently only one resolve is
implemented the local resolver.
With a manually created structure like that, one could load a
plugin:
```
repo.private
└── github.com
    └── a-peyrard
        └── addson_foobar
            └── 1.0.0
                └── arm64
                    ├── all=-N -l
                    │   └── lib.so
                    └── default
                        └── lib.so
```